GSIRelay/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bidict==0.23.1
click==8.1.7
fastapi==0.110.0
h11==0.14.0
idna==3.6
pydantic==2.6.3
pydantic_core==2.16.3
python-engineio==4.9.0
python-socketio==5.11.1
simple-websocket==1.0.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1
wsproto==1.2.0
